change: 明示的castバグ matcher の cast 元型を変更

+++ clang-query/downcast-bug/exp-downcast-bug.c
@@ -13,8 +13,8 @@ void explicitlyDowncast() {
 void downcastSttime() {
     struct stat stat_buf;
     if (stat("a.c", &stat_buf) == 0) {
-        printf("%d\n", stat_buf.st_atime);
-        printf("%d\n", stat_buf.st_atime);
+        printf("%d\n", (int)stat_buf.st_atime);
+        printf("%d\n", (int)stat_buf.st_atime);
     }
 }
